Added cObjectData class to store information about position, size, rotation, objectType, and materialID. It is used for all types of objects (fractals, primitives). ObjectId is passed between functions instead of another data like color, reflectance, etc. REMARK! Actually all primitive objects are black.

+++ mandelbulber2/src/ssao_worker.cpp
@@ -123,7 +123,7 @@ void cSSAOWorker::doWork()
 				}
 
 				double ambient = 0;
-				double angleStep = M_PI * 2.0 / quality;
+				double angleStep = M_PI * 2.0 / (double)quality;
 				double maxRandom = 62831 / quality;
 				double rRandom = 1.0;
